Set up the User profile page. Changed database schema from using AvailableTimes and Many-to-Many Users<->Services relationship to just ProvidedServices table. Darkened the background image. Moved HTML routes into the base HTML route folder. Set up files for Request View and Provided Service View.
