+++ routes/api/user/read-user.js
@@ -5,24 +5,32 @@ var db = require('../../../models');
 // Create an Express Router to allow routing via files external to server.js
 var router = express.Router();
 
-console.log('read-user');
 router.get("/api/users", function(req, res) {
-    // 1. Add a join to include all of users
-    db.Users.findAll({}).then(function(dbUser) {
-      res.json(dbUser);
-    });
-  });
+	db.Users.findAll({}).then(function(dbUser) {
+		res.json(dbUser);
+	});
+});
 
-  router.get("/api/users/:id", function(req, res) {
-    // 2; Add a join to include all of users
-    db.Users.findOne({
-      where: {
-        id: req.params.id
-      }
-    }).then(function(dbUser) {
-      res.json(dbUser);
-    });
-  });
+router.get("/api/users/:id", function(req, res) {
+	db.Users.findOne({
+		where: {
+			id: req.params.id
+		}
+	}).then(function(dbUser) {
+		res.json(dbUser);
+	});
+});
+
+// router.get('/api/users/:id/services', function (req, res) {
+// 	db.ProvidedServices.findAll({
+// 		where: {
+// 			UserId: req.params.id
+// 		}
+// 	})
+// 	.then(function (dbProvidedServices) {
+// 		res.json(dbProvidedServices);
+// 	})
+// });
 
 // Export these routers
 module.exports = router;
